pythonSDK/pyproject.toml
[project]
name = "perpex-sdk"
version = "0.1.0"
description = "Python SDK port of the Perpex TS SDK (query-first)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "web3>=7.0.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

